pyproject.toml
[project]
name = "finance-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "anthropic>=0.75.0",
    "chromadb>=1.4.0",
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.2",
    "langchain-openai>=1.1.5",
    "langgraph>=1.0.5",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "yfinance>=0.2.66",
]
